New pywb Images, Logging, Stats fixes (from production) (#686)
* misc tweaks from prod deployment:
- invalid recording detect: if recording with cdxj key missing owner, recording no longer valid, delete left over keys
- temp user cleanup: if recording still open when deleting, close first and delete on next loop
- logging: remove additional urllib logging, convert local storage to user logger
- stats: add stats for 4GB user count
- delete_file fix: return if filename is empty
- logging: add tempchecker logging
- docker: use pywb 2.2.20190227 image
- version: bump to 4.2.1

+++ webrecorder/webrecorder/rec/tempchecker.py
@@ -77,6 +77,18 @@ class TempChecker(object):
                         redis=self.data_redis,
                         access=BaseAccess())
 
+            wait_to_delete = False
+
+            for collection in user.get_collections(load=False):
+                for recording in collection.get_recordings(load=False):
+                    if recording.is_open(extend=False):
+                        recording.set_closed()
+                        logger.debug('TempChecker: Closing temp recording: ' + recording.my_id)
+                        wait_to_delete = True
+
+            if wait_to_delete:
+                return False
+
             user.delete_me()
 
             self.sesh_redis.delete(temp_key)
@@ -139,6 +151,8 @@ class TempChecker(object):
 
             if temp_user not in temps_to_remove:
                 temps_to_remove.add((temp_user, os.path.join(self.record_root_dir, temp_user)))
+
+        logger.debug('TempChecker: Temp Users to Remove: {0}'.format(len(temps_to_remove)))
 
         # remove if expired
         for temp_user, temp_dir in temps_to_remove:
